Updated Animation
Added  animations with the choice of animation 1: Light up one pixel after the other, in essence loading the whole led matrix in turquoise colour.

Animation 2: Display and blink an LED segment and gradually colour the whole LED matrix.

Animation 3: The user can select which LED segments to light up as the rows are listed in a map.